use flush_after_seconds, moved map_avro
